Separated MenuView from SimulatorView
Separated to keep SimulatorView unchanged.
