+++ src/main/java/uk/ac/aston/dc2300/ocean/world/SimulatorView.java
@@ -53,9 +53,6 @@ public class SimulatorView extends JFrame
         setLocation(100, 50);
         
         fieldView = new FieldView(height, width);
-
-        // Added menu bar to the frame
-        setJMenuBar(createMenuBar());
         
         Container contents = getContentPane();
         contents.add(stepLabel, BorderLayout.NORTH);
@@ -147,57 +144,6 @@ public class SimulatorView extends JFrame
     public boolean isViable(Field field)
     {
         return stats.isViable(field);
-    }
-    
-    public JMenuBar createMenuBar() {
-    	JMenu menu;
-        JMenuItem optionsItem;
-        JMenuItem restartSimulationItem;
-        
-        // Create the menu bar.
-        JMenuBar menuBar = new JMenuBar();
-        
-    	// Menu
-        menu = new JMenu("Menu");
-        menu.setMnemonic(KeyEvent.VK_M);
-        menu.getAccessibleContext().setAccessibleDescription(
-                "Options Menu");
-        menuBar.add(menu);
-        
-        // Options menu item
-        optionsItem = new JMenuItem("Options", KeyEvent.VK_O);
-        optionsItem.getAccessibleContext().setAccessibleDescription(
-                "Configure simulation options.");
-        menu.add(optionsItem);
-        
-        // Restart Simulation menu item
-        restartSimulationItem = new JMenuItem("Restart Simulation", KeyEvent.VK_R);
-        restartSimulationItem.getAccessibleContext().setAccessibleDescription(
-                "Restart the simulation.");
-        menu.add(restartSimulationItem);
-        
-        optionsItem.addActionListener(new ActionListener() {
-
-			@Override
-			public void actionPerformed(ActionEvent e) {
-				//TODO Open options window to set simulation constants
-				
-				// Set some dummy constants
-				ModelConstants.OCEAN_WIDTH = 50;
-				ModelConstants.OCEAN_DEPTH = 50;
-				// TODO Add dialog which tells you to restart the simulation
-			}
-		});
-        
-        restartSimulationItem.addActionListener(new ActionListener() {
-			
-			@Override
-			public void actionPerformed(ActionEvent e) {
-				dispose();
-			}
-		});
-        
-		return menuBar;
     }
     
     /**
